requirements.txt
sentencepiece>=0.1.91
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==1.8
torch>1.8.1
protobuf==3.19.*
pyyaml
transformers
accelerate
scipy
openai
matplotlib
colorama
termcolor
imageio
wandb
ipython
tqdm==4.64.0
datasets
gym==0.23.0
